fix(Player) refactored player movement code in start scene

+++ src/Engine/Events/Controls.c
@@ -147,7 +147,7 @@ bool getKeyState(SDL_Keycode code) {
  * @return True if the button is pressed, false otherwise.
  */
 bool getMouseButtonState(int button) {
-    if (button < 0 || button >= sizeof(mouseState)) {
+    if (button < 0 || button >= (int)sizeof(mouseState)) {
         logSDLError("Invalid mouse button code");
         return false;
     }
@@ -161,7 +161,7 @@ bool getMouseButtonState(int button) {
  * @return The mouse position in the specified coordinate.
  */
 int getMousePosition(int coord) {
-    if (coord < 0 || coord >= sizeof(mousePos) / sizeof(mousePos[0])) {
+    if (coord < 0 || coord >= (int)(sizeof(mousePos) / sizeof(mousePos[0]))) {
         logSDLError("Invalid mouse position coordinate");
         return -1;
     }
@@ -175,7 +175,7 @@ int getMousePosition(int coord) {
  * @param value The value to set the coordinate to.
  */
 void setMousePosition(int coord, float value) {
-    if (coord < 0 || coord >= sizeof(mousePos) / sizeof(mousePos[0])) {
+    if (coord < 0 || coord >= (int)(sizeof(mousePos) / sizeof(mousePos[0]))) {
         logSDLError("Invalid mouse position coordinate");
         return;
     }
